[chore] refractor app.
