+++ .travis.yml
@@ -1,0 +1,9 @@
+language: python
+python:
+- '3.6'
+before_script:
+- pip3 install -r requirements.txt
+script:
+- python -m pytest -v --cov app --cov-report term-missing
+after_success:
+- coveralls
