+++ app.py
@@ -1,0 +1,8 @@
+def welcome():
+    return "Welcome to Epic Rater \n"
+
+def signup_sign_menu():
+    output = "Select an option"
+    output += "\n 1. Signup"
+    output += "\n 2. Sigin"
+    return output
