+++ run.py
@@ -1,12 +1,4 @@
-def welcome():
-    return "Welcome to Epic Rater \n"
-
-def signup_sign_menu():
-    output = "Select an option"
-    output += "\n 1. Signup"
-    output += "\n 2. Sigin"
-    return output
-
+from utilities import *
 def enter_option():
     option = int(input("Your option \n"))
     if option == 1:
